Implement 'legacy' file for comparison testing - Move totally unmodified code to 2fa.legacy.js - _only_ add calls to moveSmsAndEmailContentToContext(), that we can get the message content out of routes that send messages to verify our changes are safe.
